Lab 2: Freenove Robot Code
I think you mentioned in class that we are not able to share code but, I am not sure how to do that in VNC Viewer as we share a terminal to code with for the robot. For future reference would you like us to come in separately to code portions of the robot or is the sharing code thing specifically for more home oriented projects. Thank you.